% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rk.ClusterAnalysis-package.R
\docType{package}
\name{rk.ClusterAnalysis-package}
\alias{rk.ClusterAnalysis}
\alias{rk.ClusterAnalysis-package}
\title{\packageTitle{rk.ClusterAnalysis}}
\description{
\packageDescription{rk.ClusterAnalysis}
}
\details{
The DESCRIPTION file:
\tabular{ll}{
Package: \tab rk.ClusterAnalysis\cr
Type: \tab Package\cr
Version: \tab 0.01-14\cr
Date: \tab 2015-11-19\cr
Depends: \tab rkward (>= 0.6.0)\cr
Enhances: \tab rkward\cr
License: \tab GPL (>= 3)\cr
LazyLoad: \tab yes\cr
URL: \tab https://rkward.kde.org\cr
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://rkward.kde.org}
}

}
\author{
\packageAuthor{rk.ClusterAnalysis}

Maintainer: \packageMaintainer{rk.ClusterAnalysis}
}
