% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rk.ClusterAnalysis-package.R
\docType{package}
\name{rk.ClusterAnalysis-package}
\alias{rk.ClusterAnalysis-package}
\title{The rk.ClusterAnalysis Package}
\description{
RKWard GUI to conduct k-means, model based and hierarchical cluster analyses.
}
\details{
\tabular{ll}{
Package: \tab rk.ClusterAnalysis\cr
Type: \tab Package\cr
Version: \tab 0.01-11\cr
Date: \tab 2015-03-13\cr
Depends: \tab rkward (>= 0.6.0)\cr
Enhances: \tab rkward\cr
License: \tab GPL (>= 3)\cr
LazyLoad: \tab yes\cr
URL: \tab http://rkward.sf.net\cr
}

RKWard GUI to conduct k-means, model based and hierarchical cluster analyses
}
\author{
Meik Michalke
}
\keyword{package}

