\docType{package}
\name{rk.ClusterAnalysis-package}
\alias{rk.ClusterAnalysis}
\alias{rk.ClusterAnalysis-package}
\title{The rk.ClusterAnalysis Package}
\description{
  RKWard GUI to conduct k-means, model based and
  hierarchical cluster analyses.
}
\details{
  \tabular{ll}{ Package: \tab rk.ClusterAnalysis\cr Type:
  \tab Package\cr Version: \tab 0.01-9\cr Date: \tab
  2013-04-04\cr Depends: \tab rkward (>= 0.5.6)\cr
  Enhances: \tab rkward\cr License: \tab GPL (>= 3)\cr
  LazyLoad: \tab yes\cr URL: \tab http://rkward.sf.net\cr }

  RKWard GUI to conduct k-means, model based and
  hierarchical cluster analyses
}
\author{
  Meik Michalke \email{meik.michalke@hhu.de}
}
\keyword{package}

